Update (#8)
+ `ArrayPoolExtensions` for renting a cleared array
+ `RentedArray<T>` supports clearing the rented array on construction
+ `RentedArray<T>` supports clearing the rented array on disposing by setting `Clear` to `true`
- Fixed `long.RemoveFlags` extension `flags` parameter type

+++ src/Wan24-Core/BitwiseInt64Extensions.cs
@@ -43,7 +43,7 @@
         /// <param name="value">Value</param>
         /// <param name="flags">Flags</param>
         /// <returns>Value</returns>
-        public static long RemoveFlags(this long value, int flags) => value & ~flags;
+        public static long RemoveFlags(this long value, long flags) => value & ~flags;
 
         /// <summary>
         /// Cast as signed byte
